client IO API update
