+++ projects/ClientQtGraphicAPI/src/qt_input_output_api/IO_disk.hpp
@@ -27,7 +27,8 @@
 
 #include "utils/fileutils.hpp"
 #include "utils/log.hpp"
-#include "client_redemption/client_redemption_api.hpp"
+//#include "client_redemption/client_redemption_api.hpp"
+#include "client_redemption/client_input_output_api/client_iodisk_api.hpp"
 #include "core/RDP/channels/rdpdr.hpp"
