+++ projects/ClientQtGraphicAPI/src/qt_input_output_api/qt_IO_graphic_mouse_keyboard.hpp
@@ -26,14 +26,16 @@
 
 
 
-#include "client_redemption/client_redemption_api.hpp"
+// #include "client_redemption/client_redemption_api.hpp"
 #include "keymaps/qt_scancode_keymap.hpp"
+
 #include "qt_graphics_components/qt_progress_bar_window.hpp"
 #include "qt_graphics_components/qt_options_window.hpp"
 #include "qt_graphics_components/qt_screen_window.hpp"
 #include "qt_graphics_components/qt_form_window.hpp"
 
 #include "client_redemption/client_input_output_api/client_graphic_api.hpp"
+#include "client_redemption/client_input_output_api/client_mouse_keyboard_api.hpp"
 
 #include <QtGui/QBitmap>
 #include <QtGui/QColor>
