+++ projects/ClientQtGraphicAPI/src/qt_input_output_api/qt_graphics_components/qt_options_window.hpp
@@ -29,7 +29,7 @@
 #include "core/RDP/MonitorLayoutPDU.hpp"
 #include "core/channel_list.hpp"
 #include "client_redemption/client_redemption_api.hpp"
-
+#include "client_redemption/client_input_output_api/client_mouse_keyboard_api.hpp"
 
 
 #include "../keymaps/qt_scancode_keymap.hpp"
