+++ projects/ClientQtGraphicAPI/src/qt_input_output_api/qt_graphics_components/qt_screen_window.hpp
@@ -25,7 +25,7 @@
 #include "utils/log.hpp"
 #include "client_redemption/client_redemption_api.hpp"
 #include "client_redemption/client_input_output_api/client_graphic_api.hpp"
-
+#include "client_redemption/client_input_output_api/client_mouse_keyboard_api.hpp"
 
 #include <QtCore/QTimer>
 #include <QtGui/QMouseEvent>
