+++ projects/ClientQtGraphicAPI/src/qt_input_output_api/qt_input_output_clipboard.hpp
@@ -31,6 +31,7 @@
 #include "utils/fileutils.hpp"
 
 #include "client_redemption/client_redemption_api.hpp"
+#include "client_redemption/client_input_output_api/client_clipboard_api.hpp"
 
 #include <QtCore/QMimeData>
 #include <QtGui/QClipboard>
