+++ projects/ClientQtGraphicAPI/src/qt_input_output_api/qt_input_socket.hpp
@@ -25,7 +25,8 @@
 
 #include "utils/log.hpp"
 #include "core/session_reactor.hpp"
-#include "client_redemption/client_redemption_api.hpp"
+//#include "client_redemption/client_redemption_api.hpp"
+#include "client_redemption/client_input_output_api/client_socket_api.hpp"
 
 #if REDEMPTION_QT_VERSION == 4
 #   define REDEMPTION_QT_INCLUDE_WIDGET(name) <QtGui/name>
