+++ projects/ClientQtGraphicAPI/src/qt_input_output_api/qt_output_sound.hpp
@@ -24,7 +24,7 @@
 #include <fstream>
 
 #include "utils/log.hpp"
-#include "client_redemption/client_redemption_api.hpp"
+#include "client_redemption/client_input_output_api/client_sound_api.hpp"
 
 #if REDEMPTION_QT_VERSION == 4
 #   define REDEMPTION_QT_INCLUDE_WIDGET(name) <QtGui/name>
