+++ src/client_redemption/client_channel_managers/client_channel_CLIPRDR_manager.hpp
@@ -25,7 +25,7 @@
 
 #include "utils/log.hpp"
 #include "client_redemption/client_redemption_api.hpp"
-
+#include "client_redemption/client_input_output_api/client_clipboard_api.hpp"
 #include "core/RDP/clipboard.hpp"
