+++ src/client_redemption/client_channel_managers/client_channel_RDPDR_manager.hpp
@@ -34,7 +34,8 @@
 #include "core/FSCC/FileInformation.hpp"
 #include "core/RDP/channels/rdpdr.hpp"
 
-#include "client_redemption/client_redemption_api.hpp"
+//#include "client_redemption/client_redemption_api.hpp"
+#include "client_redemption/client_input_output_api/client_iodisk_api.hpp"
